Parametros generales CRUD

+++ IICAPS v1/DataObject/CreditoAlumno.cs
@@ -10,12 +10,12 @@ namespace IICAPS_v1.DataObject
     {
         public int id { get; set; }
         public string alumno { get; set; }
-        public double cantidadMensualidad { get; set; }
+        public decimal cantidadMensualidad { get; set; }
         public int cantidadMeses { get; set; }
         public DateTime fechaSolicitud { get; set; }
         public string observaciones { get; set; }
         public string estado { get; set; }
-        public double cantidadAbonoCredito { get; set; }
-        public double cantidadAbonoMensual { get; set; }
+        public decimal cantidadAbonoCredito { get; set; }
+        public decimal cantidadAbonoMensual { get; set; }
     }
 }
